backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
pydantic
python-dotenv
tiktoken
pypdf
